@@ -1,10 +1,10 @@
 name: color_picker_example
 description: Two examples of how to use the FlexColorPicker package.
-version: 3.6.0
+version: 3.7.0
 publish_to: 'none'
 environment:
   sdk: '>=2.19.0 <4.0.0'
-  flutter: '>=3.22.0'
+  flutter: '>=3.27.0'
 
 dependencies:
   # https://pub.dev/packages/flex_color_picker
