@@ -702,7 +702,7 @@ class ColorPicker extends StatefulWidget {
   /// Defaults to false.
   final bool colorCodeReadOnly;
 
-  /// Set to true to show the int [Color.value] of the selected [color].
+  /// Set to true to show the int [Color] value of the selected [color].
   ///
   /// This is a developer feature, showing the int color value can be
   /// useful during software development. If enabled the value is shown after
@@ -1469,7 +1469,7 @@ class _ColorPickerState extends State<ColorPicker> {
     _selectedColor = widget.color.withAlpha(0xFF);
     _tappedColor = widget.color;
     // Opacity is captured in _opacity if enabled.
-    _opacity = widget.enableOpacity ? widget.color.opacity : 1;
+    _opacity = widget.enableOpacity ? widget.color.a : 1;
     // Picker labels, use english fallbacks if none provided.
     _pickerLabels = <ColorPickerType, String>{
       ColorPickerType.both: widget.pickerTypeLabels[ColorPickerType.both] ??
@@ -1558,7 +1558,7 @@ class _ColorPickerState extends State<ColorPicker> {
     bool shouldFindPickerAndSwatch = false;
     // Opacity enable/disable changed, update selected color and opacity.
     if (widget.enableOpacity != oldWidget.enableOpacity) {
-      _opacity = widget.enableOpacity ? widget.color.opacity : 1;
+      _opacity = widget.enableOpacity ? widget.color.a : 1;
       if (_debug) {
         debugPrint('didUpdateWidget changed: enableOpacity = '
             '${widget.enableOpacity == oldWidget.enableOpacity}'
@@ -1572,7 +1572,7 @@ class _ColorPickerState extends State<ColorPicker> {
             'color=${widget.color} selectedColor=$_selectedColor');
       }
       _selectedColor = widget.color.withAlpha(0xFF);
-      _opacity = widget.enableOpacity ? widget.color.opacity : 1;
+      _opacity = widget.enableOpacity ? widget.color.a : 1;
       // Make a swatch too be to find it on wheel, if color is there.
       _typeToSwatchMap[ColorPickerType.wheel] = <ColorSwatch<Object>>[
         ColorTools.createPrimarySwatch(_selectedColor),
@@ -2045,8 +2045,8 @@ class _ColorPickerState extends State<ColorPicker> {
                     shouldRequestsFocus: _wheelShouldFocus,
                     onChangeStart: (Color color) {
                       widget.onColorChangeStart
-                          ?.call(color.withOpacity(_opacity));
-                      _addToRecentColors(color.withOpacity(_opacity));
+                          ?.call(color.withValues(alpha: _opacity));
+                      _addToRecentColors(color.withValues(alpha: _opacity));
                     },
                     onChanged: (Color color) {
                       setState(() {
@@ -2061,12 +2061,12 @@ class _ColorPickerState extends State<ColorPicker> {
                         _wheelShouldFocus = false;
                         _updateActiveSwatch();
                       });
-                      widget
-                          .onColorChanged(_selectedColor.withOpacity(_opacity));
+                      widget.onColorChanged(
+                          _selectedColor.withValues(alpha: _opacity));
                     },
                     onChangeEnd: (Color color) {
                       widget.onColorChangeEnd?.call(
-                        color.withOpacity(_opacity),
+                        color.withValues(alpha: _opacity),
                       );
                     },
                     onWheel: (bool value) {
@@ -2191,9 +2191,9 @@ class _ColorPickerState extends State<ColorPicker> {
                             _opacity = value;
                           });
                           widget.onColorChangeStart!(
-                              _selectedColor.withOpacity(_opacity));
+                              _selectedColor.withValues(alpha: _opacity));
                           _addToRecentColors(
-                              _selectedColor.withOpacity(_opacity));
+                              _selectedColor.withValues(alpha: _opacity));
                         }
                       },
                       onChanged: (double value) {
@@ -2206,7 +2206,7 @@ class _ColorPickerState extends State<ColorPicker> {
                           _wheelShouldFocus = false;
                         });
                         widget.onColorChanged(
-                            _selectedColor.withOpacity(_opacity));
+                            _selectedColor.withValues(alpha: _opacity));
                       },
                       onChangeEnd: (double value) {
                         if (widget.onColorChangeEnd != null) {
@@ -2215,7 +2215,7 @@ class _ColorPickerState extends State<ColorPicker> {
                             _opacity = value;
                           });
                           widget.onColorChangeEnd!(
-                              _selectedColor.withOpacity(_opacity));
+                              _selectedColor.withValues(alpha: _opacity));
                           // _addToRecentColors(
                           //     _selectedColor.withOpacity(_opacity));
                         }
@@ -2299,7 +2299,7 @@ class _ColorPickerState extends State<ColorPicker> {
                       if (widget.showColorCode)
                         ColorCodeField(
                           color: widget.enableOpacity
-                              ? _selectedColor.withOpacity(_opacity)
+                              ? _selectedColor.withValues(alpha: _opacity)
                               : _tappedColor,
                           readOnly: _activePicker != ColorPickerType.wheel ||
                               widget.colorCodeReadOnly,
@@ -2311,8 +2311,8 @@ class _ColorPickerState extends State<ColorPicker> {
                           enableTooltips: widget.enableTooltips,
                           shouldUpdate: _editShouldUpdate,
                           onColorChanged: (Color color) {
-                            widget.onColorChangeStart
-                                ?.call(_selectedColor.withOpacity(_opacity));
+                            widget.onColorChangeStart?.call(
+                                _selectedColor.withValues(alpha: _opacity));
                             setState(() {
                               _tappedColor = color;
                               _selectedColor = color;
@@ -2324,11 +2324,11 @@ class _ColorPickerState extends State<ColorPicker> {
                               _updateActiveSwatch();
                             });
                             widget.onColorChanged(
-                                _selectedColor.withOpacity(_opacity));
-                            widget.onColorChangeEnd
-                                ?.call(_selectedColor.withOpacity(_opacity));
+                                _selectedColor.withValues(alpha: _opacity));
+                            widget.onColorChangeEnd?.call(
+                                _selectedColor.withValues(alpha: _opacity));
                             _addToRecentColors(
-                                _selectedColor.withOpacity(_opacity));
+                                _selectedColor.withValues(alpha: _opacity));
                           },
                           onEditFocused: (bool editInFocus) {
                             _requestEditFocus = false;
@@ -2349,7 +2349,7 @@ class _ColorPickerState extends State<ColorPicker> {
                         const SizedBox(width: 8),
                       if (widget.showColorValue)
                         SelectableText(
-                          _selectedColor.value.toString(),
+                          _selectedColor.value32bit.toString(),
                           style: effectiveCodeStyle,
                         ),
                     ],
@@ -2406,12 +2406,13 @@ class _ColorPickerState extends State<ColorPicker> {
     // uses this flag. If we use Recent colors, we should also extract the
     // opacity from the selected color, and set it as the current opacity.
     if (findPicker) {
-      _opacity = color.opacity;
+      _opacity = color.a;
     }
     // Call start callback with current selectedColor before change.
     if (widget.enableOpacity) {
-      widget.onColorChangeStart?.call(_selectedColor.withOpacity(_opacity));
-      _addToRecentColors(_selectedColor.withOpacity(_opacity));
+      widget.onColorChangeStart
+          ?.call(_selectedColor.withValues(alpha: _opacity));
+      _addToRecentColors(_selectedColor.withValues(alpha: _opacity));
     } else {
       // This is to allow custom colors with opacity to callback with their
       // in-built opacity, when opacity is control is not enabled.
@@ -2463,8 +2464,8 @@ class _ColorPickerState extends State<ColorPicker> {
     });
     // Call the change call back with the new color.
     if (widget.enableOpacity) {
-      widget.onColorChanged(_selectedColor.withOpacity(_opacity));
-      widget.onColorChangeEnd?.call(_selectedColor.withOpacity(_opacity));
+      widget.onColorChanged(_selectedColor.withValues(alpha: _opacity));
+      widget.onColorChangeEnd?.call(_selectedColor.withValues(alpha: _opacity));
     } else {
       // This is to allow custom colors with opacity to callback with their
       // in-built opacity, when opacity is control is not enabled.
@@ -2520,8 +2521,8 @@ class _ColorPickerState extends State<ColorPicker> {
         // We always remove any alpha from pasted colors.
         _selectedColor = clipColor.withAlpha(0xFF);
         // If opacity is enabled, we capture the opacity from the pasted color.
-        _opacity = widget.enableOpacity ? clipColor.opacity : 1;
-        _addToRecentColors(_selectedColor.withOpacity(_opacity));
+        _opacity = widget.enableOpacity ? clipColor.a : 1;
+        _addToRecentColors(_selectedColor.withValues(alpha: _opacity));
         // Color changed outside wheel and edit field, a new shade or color was
         // selected outside the wheel and edit, they should update!
         _wheelShouldUpdate = true;
