@@ -1,4 +1,5 @@
 import 'package:flex_color_picker/flex_color_picker.dart';
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 
 /// Contains used constants for doubles, strings and colors.
@@ -9,14 +10,24 @@ class App {
   // Web demo with inside the app. Shown on the start screen in the demo,
   // so people testing it don't have to ask. Also info for the About screen.
   static const String appName = 'FlexColorPicker';
-  static const String version = '3.6.0';
+  static const String version = '3.7.0';
   static const String packageVersion = 'FlexColorPicker package $version';
   static final Uri packageUri = Uri(
     scheme: 'https',
     host: 'pub.dev',
     path: 'packages/flex_color_picker',
   );
-  static const String flutterVersion = 'Channel stable 3.24.3 (canvaskit)';
+
+  // Check if this is a Web-WASM build, Web-JS build or native VM build.
+  static const bool isRunningWithWasm =
+      bool.fromEnvironment('dart.tool.dart2wasm');
+  static const String buildType = isRunningWithWasm
+      ? ', WasmGC'
+      : kIsWeb
+          ? ', JS'
+          : ', native VM';
+
+  static const String flutterVersion = 'stable 3.27.0 (canvaskit$buildType)';
   static const String copyright = '© 2020 - 2024';
   static const String author = 'Mike Rydstrom';
   static const String license = 'BSD 3-Clause License';
