@@ -1,6 +1,6 @@
 /// A customizable Flutter primary, accent and custom color picker. Includes
 /// an optional HSV wheel color picker.
-library flex_color_picker;
+library;
 
 export 'src/color_indicator.dart' show ColorIndicator;
 export 'src/color_picker.dart' show ColorPicker, showColorPickerDialog;
