@@ -445,7 +445,7 @@ Future<Color> showColorPickerDialog(
   /// Defaults to false.
   bool colorCodeReadOnly = false,
 
-  /// Set to true to show the int [Color.value] of the selected `color`.
+  /// Set to true to show the int [Color] value of the selected `color`.
   ///
   /// This is a developer feature, showing the int color value can be
   /// useful during software development. If enabled the value is shown after
