@@ -85,8 +85,8 @@ class RecentColors extends StatelessWidget {
       children: <Widget>[
         for (final Color color in recentColors)
           ColorIndicator(
-            isSelected:
-                selectedColor == color || selectedColor.value == color.value,
+            isSelected: selectedColor == color ||
+                selectedColor.value32bit == color.value32bit,
             color: color,
             width: width,
             height: height,
